+++ lua/sqlserver/default_opts.lua
@@ -49,6 +49,7 @@ return {
   timeouts = {
     lsp_attach = 10000,
     connection = 10000,
+    export = 10000,
     object_explorer = 10000,
     query = false,
   },